+++ assets/src/js/frontend/give.js
@@ -20,7 +20,7 @@ import iFrameResizer from '../plugins/form-template/iframe-content';
 import '../plugins/form-template/parent-page';
 import '../../../../src/Form/LegacyConsumer/resources/js/conditional-fields';
 
-import '../../../../src/DonationForms/resources/utils/asyncLoadDataOnFormListViews';
+import '../../../../src/DonationForms/resources/utils/loadAsyncDataOnFormListViews';
 
 window.addEventListener('load', function () {
     window.Give.WINDOW_IS_LOADED = true;
